win-requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
comtypes==1.1.11
cryptography==36.0.2
cycler==0.11.0
easydict==1.9
idna==3.3
kiwisolver==1.4.0
matplotlib==3.5.1
my-voice-analysis==0.7
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
praat-parselmouth==0.4.1
pycochleagram==0.1 // nueiti i aplanakala per terminala ir parasyti python setup.py install
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pywin32==302
pywinauto==0.6.8
requests==2.27.1
scipy==1.8.0
six==1.16.0
SoundFile==0.10.3.post1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.9

